# Conflicts:
#	sparkzxl-alarm-starter/pom.xml
#	sparkzxl-boot-starter/pom.xml
#	sparkzxl-cache-starter/pom.xml
#	sparkzxl-core/pom.xml
#	sparkzxl-datasource-starter/pom.xml
#	sparkzxl-drools-starter/pom.xml
#	sparkzxl-elasticsearch-starter/pom.xml
#	sparkzxl-entity/pom.xml
#	sparkzxl-job-executor-starter/pom.xml
#	sparkzxl-jwt-starter/pom.xml
#	sparkzxl-log-starter/pom.xml
#	sparkzxl-mail-starter/pom.xml
#	sparkzxl-mongodb-starter/pom.xml
#	sparkzxl-mybatis-starter/pom.xml
#	sparkzxl-oauth2-resource-starter/pom.xml
#	sparkzxl-oss-starter/pom.xml
#	sparkzxl-patterns-starter/pom.xml
#	sparkzxl-pdf-starter/pom.xml
#	sparkzxl-security-starter/pom.xml
#	sparkzxl-sentinel-starter/pom.xml
#	sparkzxl-sms-starter/pom.xml
#	sparkzxl-swagger-starter/pom.xml
#	sparkzxl-uid/pom.xml
#	sparkzxl-user-starter/pom.xml
#	sparkzxl-web-starter/pom.xml

+++ sparkzxl-sms-starter/src/main/java/com/github/sparkzxl/sms/constant/enums/SendStatusEnum.java
@@ -2,26 +2,24 @@ package com.github.sparkzxl.sms.constant.enums;
 
 import lombok.Getter;
 import lombok.RequiredArgsConstructor;
-import org.apache.commons.lang3.StringUtils;
-import org.springframework.util.ObjectUtils;
 
 /**
  * description: 短信发送状态枚举
  *
  * @author zhouxinlei
- * @date 2021-12-28 10:18:10
+ * @since 2021-12-28 10:18:10
  */
 @Getter
 @RequiredArgsConstructor
 public enum SendStatusEnum {
 
-    WAITING(1L, "WAITING"),
-    FAILURE(2L, "FAIL"),
-    SUCCESS(3L, "SUCCESS");
-    private final Long code;
+    WAITING(1, "WAITING"),
+    FAILURE(2, "FAIL"),
+    SUCCESS(3, "SUCCESS");
+    private final Integer code;
     private final String desc;
 
-    public static SendStatusEnum get(Long code) {
+    public static SendStatusEnum get(Integer code) {
         for (SendStatusEnum sendStatusEnum : SendStatusEnum.values()) {
             if (sendStatusEnum.getCode() == code) {
                 return sendStatusEnum;
